requirements.txt
streamlit>=1.28.0
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
pandas>=2.0.0
plotly>=5.0.0
db-dtypes>=1.0.0
openpyxl>=3.0.0
pyarrow>=10.0.0
